+++ Toggl.Giskard/Views/ReportsRecyclerViewHolder.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using Android.Content;
-using Android.Graphics;
 using Android.Runtime;
 using Android.Views;
 using MvvmCross.Binding.Droid.BindingContext;
@@ -20,8 +18,8 @@ namespace Toggl.Giskard.Views
         {
             var context = Mvx.Resolve<IMvxAndroidGlobals>().ApplicationContext;
 
-            lastContainerHeight = (int)72.DpToPixels(context);
-            normalContainerHeight = (int)48.DpToPixels(context);
+            lastContainerHeight = 72.DpToPixels(context);
+            normalContainerHeight = 48.DpToPixels(context);
         }
 
         public bool IsLastItem { get; set; }
